OK - v5

+++ src/users/users.controller.ts
@@ -1,8 +1,12 @@
 import {
+  BadRequestException,
   Body,
   Controller,
   Delete,
   Get,
+  HttpCode,
+  Inject,
+  NotFoundException,
   Param,
   Post,
   Query,
@@ -11,7 +15,9 @@ import {
 import { UsersService } from './users.service';
 import { IsEmail, Length } from 'class-validator';
 import { AuthBaseGuard } from '../guards/AuthBase.guard';
-import { ApiTags } from '@nestjs/swagger';
+import { ApiBasicAuth, ApiTags } from '@nestjs/swagger';
+import { Model } from 'mongoose';
+import { UsersSchemaInterface } from './users.schemas';
 
 export class BodyCreateUserType {
   @Length(3, 10)
@@ -24,22 +30,64 @@ export class BodyCreateUserType {
 @ApiTags('users')
 @Controller('users')
 export class UsersController {
-  constructor(protected usersService: UsersService) {}
+  constructor(
+    protected usersService: UsersService,
+    @Inject('USERS_MODEL')
+    private usersRepository: Model<UsersSchemaInterface>,
+  ) {}
   @Get()
+  @ApiBasicAuth()
+  @UseGuards(AuthBaseGuard)
   getUsers(
     @Query('pageNumber') pageNumber: string,
     @Query('pageSize') pageSize: string,
+    @Query('sortBy') sortBy: string,
+    @Query('sortDirection') sortDirection: string,
+    @Query('searchLoginTerm') searchLoginTerm: string,
+    @Query('searchEmailTerm') searchEmailTerm: string,
   ) {
-    return this.usersService.getUsers(pageNumber, pageSize);
+    return this.usersService.getUsers(
+      pageNumber,
+      pageSize,
+      sortBy,
+      sortDirection,
+      searchLoginTerm,
+      searchEmailTerm,
+    );
   }
-  @UseGuards(AuthBaseGuard)
   @Post()
-  createUser(@Body() bodyCreateUser: BodyCreateUserType) {
+  @ApiBasicAuth()
+  @UseGuards(AuthBaseGuard)
+  async createUser(@Body() bodyCreateUser: BodyCreateUserType) {
+    const login = await this.usersRepository.findOne({
+      'accountData.login': bodyCreateUser.login,
+    });
+    const email = await this.usersRepository.findOne({
+      'accountData.email': bodyCreateUser.email,
+    });
+    if (login || email) {
+      const error = [];
+      if (login) {
+        error.push('login is repeated');
+      }
+      if (email) {
+        error.push('email is repeated');
+      }
+      throw new BadRequestException({ message: error });
+    }
     return this.usersService.createUser(bodyCreateUser);
   }
-  @UseGuards(AuthBaseGuard)
   @Delete(':id')
-  deleteUser(@Param('id') deleteUser: string) {
+  @UseGuards(AuthBaseGuard)
+  @ApiBasicAuth()
+  @HttpCode(204)
+  async deleteUser(@Param('id') deleteUser: string) {
+    const userId = await this.usersRepository.findOne({
+      'accountData.userId': deleteUser,
+    });
+    if (!userId) {
+      throw new NotFoundException('not found');
+    }
     return this.usersService.deleteUser(deleteUser);
   }
 }
